Fixed data processing logic, split code into files
